finally, wit is working reasonably well.

+++ config/initializers/wit_ai.rb
@@ -1,9 +1,5 @@
 require 'wit'
 
-access_token = ENV['WIT_ACCESS_TOKEN']
-
-
-
 # note, context: it's a hash with string keys. allways string keys
 # that's because we save it in json in redis for reuse
 actions = {
@@ -20,7 +16,8 @@ actions = {
     # wit relies on us to know what contexts and entites result in
     # entirely new contexts and entities.
 
-    # this is such a hack
+    # this whole thing is such a hack
+    # needs to be a state machine. for sure.
 
     context.merge!(entities) # how we know what happened
 
@@ -28,64 +25,88 @@ actions = {
     pp context
     puts "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
     # handle initial yes or no
-    if context['want_interview'].nil? && context['yes_no']
+
+    # # when the person sends an unintelligible date
+    if context['want_interview'] && context['datetime'].nil?
+      puts "want an interview, but no date"
+      # they may have previously sent a good date.
+      context.delete('date_is_valid') if context['date_is_valid']
+      context['date_is_invalid'] = true
+    end
+
+    if context['want_interview'].nil? && context['yes_no'] && context['reference_time_slot']
+      puts "first 'yes or no'"
       case context['yes_no'][0]['value']
       when 'yes'
-        context.merge!({ 'want_interview' => true })
+        context.delete('refuse_interview')
+        context['want_interview'] = true
       when 'no'
-        context.merge!({ 'refuse_interview' => true })
+        context['refuse_interview'] = true
       end
       context.delete('yes_no') # start afresh
     end
 
     # handle confirming date
     if context['date_is_valid'] && context['yes_no']
+      puts "confirming date yes_no"
+
+      context.delete('reference_time_slot')
+      context.delete('want_interview')
+      context.delete('human_date')
+      context.delete('date_is_valid')
       case context['yes_no'][0]['value']
       when 'yes'
         context.delete('date_not_confirmed') # might be second attempt
-        context.merge!({ 'date_confirmed' => true })
+        context['date_confirmed'] = true
       when 'no'
         context.delete('datetime')
-        context.delete('date_is_valid')
-        context.delete('human_date')
-        context.merge!({ 'date_not_confirmed' => true })
+        context.delete('date_confirmed') # not possible, I think.
+        context['date_not_confirmed'] = true
       end
-      context.delete('yes_no') # start afresh
+      context.delete('yes_no') # start afresh, maybe?
     end
+
+    # case when we have a valid date
+    if context['datetime'] # should we check it's an interval?
+      puts "we have a datetime!"
+      context.delete('reference_time_slot')
+      context.delete('want_interview')
+      context.delete('date_is_invalid')
+      context.delete('date_not_confirmed')
+      context['date_is_valid'] = true
+      # we add human date in a function below
+    end
+
+
 
     # when the person sends a command of any sort
     if context['command']
       case context['command'][0]['value']
       when 'decline'
+        context.delete('event_id')
         context.delete('want_interview') # can decline at any time
-        context.merge!({ 'refuse_interview' => true })
+        context['decline'] = true
       end
       context.delete('command') # start afresh
     end
 
-    # case when we have a valid date
-    if context['datetime'] # should we check it's an interval?
-      context.delete('date_is_invalid')
-      context.merge!({'date_is_valid'=>true})
-    end
 
     # might not need this.
     if context['date_not_confirmed']
+
       # must remove this context if we get a new datetime
       # currently, if a person doesn't confirm the date, we enter into a loop.
       # how do I know that this is a persons second attempt to add a datetime?
     end
 
 
-    # # when the person sends an unintelligible date
-    # if context['want_interview'] && context['datetime'].nil?
-    #   # they may have previously sent a good date.
-    #   context.delete('date_is_valid') if context['date_is_valid']
-    #   context.merge!({'date_is_invalid' => true})
-    # end
+    # handling locking for event. extend lock!
+    if context['event_id'] && Redis.current.get("event_lock:#{context['person_id']}") == context['event_id']
+      Redis.current.setex("event_lock:#{context['person_id']}",3600,context['event_id'])
+    end
 
+    # context gets set in merge
     Redis.current.set("wit_context:#{context['person_id']}", context.to_json)
-    Redis.current.expire("wit_context:#{context['person_id']}", 3600)
     puts "after merge"
     pp context
     puts "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
@@ -131,16 +152,26 @@ actions = {
         selected_time = r.start_datetime_human
         msg = "A #{duration} minute conversation has been booked for:\n#{selected_time}\nWith: #{r.user.name}, \nTel: #{r.user.phone_number}\n.You'll get a reminder that morning."
       else
-        pp r.errors
+        pp r.errors if r
         msg = 'It appears there are no more times available. There will be more opportunities soon!'
       end
-      Redis.current.del("wit_context:#{person.id}")
-      context.merge!({ 'reservation_msg' => msg })
+
+
+      # cleanup context
+      %w(want_interview yes_no reference_time_slot event_id reference_time datetime date_confirmed date_is_valid).each  do |i|
+        context.delete(i)
+      end
+
+      Redis.current.set("wit_context:#{person.id}",context.to_json)
+
+      # remove our lock
+      Redis.current.del("event_lock:#{person.id}")
+      context['reservation_msg'] = msg
     else
       # lets ask for the date again
-      context.merge!({ 'date_is_invalid' => true })
+      context['date_is_invalid'] = true
     end
-    context
+    return context
   end,
   confirm_reservation: lambda do |_session_id, context|
     # person confirms reservation
@@ -158,29 +189,31 @@ actions = {
     # a person's calendar for the next few days
     context
   end,
-  date_not_confirmed: lambda do |_session_id, context|
-    # a person's calendar for the next few days
-    context.merge!({'date_not_confirmed'=>true})
-  end,
-  date_confirmed: lambda do |_session_id, context|
-    # a person's calendar for the next few days
-    context.merge!({'date_confirmed'=>true})
-  end,
   decline_invitation: lambda do |_session_id, context|
     puts "in decline"
     pp context
     puts "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
-    context.delete('want_interview') if context('want_interview')
-    # probably destroy the context and turn end the current session
-    Redis.current.del("wit_context:#{context['person_id']}")
-    context.merge('decline_invitation'=>true)
+    # cleanup our context
+
+    context.delete('want_interview')
+    context.delete('refuse_interview')
+    context.delete('reference_time_slot')
+    context.delete('reference_time')
+    context.delete('event_id')
+    context.delete('yes_no')
+    context.delete('datetime')
+    context.delete('command') # possible that the decline came from a command
+
+    # destroy the lock, should notify alexis too
+    Redis.current.del("event_lock:#{context['person_id']}")
+    Redis.current.set("wit_context:#{context['person_id']}",context.to_json)
+    return context
   end,
   date_to_human: lambda do |_session_id, context|
 
     puts "in date_to_human"
     pp context
     puts "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
-    context.delete('yes_no')
     ei = V2::EventInvitation.find_by(id: context['event_id'])
     times = get_times(context,ei.end_datetime)
     if times
@@ -197,9 +230,10 @@ actions = {
       puts "why didn't we get times here?"
       context.delete('human_date')
       context.delete('date_is_valid')
-      context.merge!({ 'date_is_invalid' => true })
+      context['date_is_invalid'] = true
     end
-    context
+    Redis.current.set("wit_context:#{context['person_id']}",context.to_json)
+    return context
   end,
 }
 
@@ -227,12 +261,14 @@ def get_times(context, default_end)
       # wit.ai uses the end of a duration if grain = hour.
       # we want the start.
       # for them 1pm to 5pm == 1pm untill 6pm, including the whole of 5pm
-      untill -= 1.hour if datetime['grain'] == 'hour'
+      untill -= 1.hour if datetime['to']['grain'] == 'hour'
+      untill -= 1.minute if datetime['to']['grain'] == 'minute'
     end
     res << { from: from, untill: untill, confidence: datetime['confidence'] }
   end
   res # not yet ready to deal with multiple time windows
 end
+
 
 
 # key contexts:
@@ -241,5 +277,6 @@ end
 # reference_time_slot: string of the start and end time of event
 # reservation_msg: either you got a slot in your time or you didn't
 # human_date: a human array of dates, joined by 'and'
+access_token = ENV['WIT_ACCESS_TOKEN']
 ::WitClient = Wit.new access_token, actions
